requirements.txt
Flask==2.2.5
pyodbc==4.0.34
azure-storage-blob==12.20.0
requests==2.31.0
azure-storage-queue>=12.0.0,<13.0.0
azure-functions
opencensus-ext-azure
opencensus-ext-flask
